HACK: Load additional config based on hardcoded pin
This change is made for Fablab Amersfoort, to allow loading a different
config when a rotational axis has been connected. This hardcodes the pin
used for that (1.29 == ZMAX), since any dynamic method would get
complicated and non-elegant due to the mixing of loading and
interpreting config. This, and the additional memory pressure would
probably prevent merging such a change upstream, and if we need a
customized version anyway, we might as well keep it simple by hardcoding
this pin.