HACK: Allow swapping y and z axes
This adds an option to the CartesianSolution to swap these axes.
However, it turns out that the Endstops module makes assumptions about
this mapping, so even when you modify the config to home Z instead of Y,
it internaly still maps Z to gamma instead of beta. So this lets the
Endstops module look at the same swap configuration value and swap the
actuators back...